feat: add .env.example, .gitignore, and docker-compose.yaml files
Add .env.example file to provide an example of the required environment variable SA_PASSWORD.

Add .gitignore file to ignore the data directory and .env file.

Add docker-compose.yaml file to define the database service using the latest version of the MSSQL server image. Set the SA_PASSWORD environment variable, accept the EULA, and specify the MSSQL_PID as Express. Expose port 5434 on the host machine and mount the data directory to the container.